Solve day 3 part 1
My solution is basically to iterate through each string and use a map as
a pseudo-set to track when elements are in the first half of the string.
Then in the second half of the string, the duplicate element can be
identified by checked whether each element is already in the set.