2022/5: Solve part 2
To keep the crates in the same order during a move it is necessary to
use an intermediate stack to push the moved crates on to. Once all
crates have been 'picked up' by the intermediate stack they can be
popped from it and pushed onto the final stack. Another approach would
be to implement a special pop() method on the stacks that can pop
multiple crates in order.