2022/6: Solve part 1
The idea with this solution is to keep a sliding window that is the size
of the sequence of unique characters we're looking for, and a map to
quickly determine whether the characters in the window are unique. When
characters are evicted from the window they are also evicted from the
map so when the length of the map equals the capacity, all the
characters in the window are unique.