2022/6: Solve part 2
Somehow I still got the correct answer for part 1 despite my
implementation not being correct. Rather than caching each rune in the
window with a boolean, it is necessary to keep track of how many times
each rune appears in the window. With that fixed the solution to part 2
is simply to change the length of the string of unique
characters.